Updated planets to have an area of effect instead of only being triggered when no other gravity source is present. This helps for situations where the player needs to walk off of platform cubes and back onto the planet.
Removed the gravity source from the planet surface as it is now redundant. Changed god script to assign with GetComponentInChildren to compensate for this change.

Moved the if null check in the attractedObject script from Awake to Start because items (like the beetles) that belong to a pool are created before the default gravity source and this caused null reference exceptions.